requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.1
